+++ cmd/opentelemetry-collector/app/exporter/grpcplugin/testdata/jaeger-config.yaml
